Move ProductCompact class to app/store/product.py
