+++ app/templates/components/order-confirm.html
@@ -1,11 +1,9 @@
 {% macro order_confirm(session_id, products, count, total_price, placement_status) %}
   <div class="relative z-10" aria-labelledby="confirm-dialog" role="dialog" aria-modal="true">
-    
     <!-- background -->
     <div class="fixed inset-0 bg-gray-500 bg-opacity-75 transition-opacity" aria-hidden="true"></div>
     <div class="fixed inset-0 z-10 w-screen">
       <div class="flex h-full items-end justify-center py-4 px-10 text-center">
-
         <!-- dialog -->
         <div class="flex flex-col overflow-y-auto transform overflow-hidden rounded-lg bg-white text-left shadow-xl transition-all max-h-full">
           <!-- text area wrapper -->
